Fix platforms where `c_ulong` is 32 bits wide

+++ src-tauri/src/platform/shared.rs
@@ -19,7 +19,7 @@ pub const IGNORE_AREA_BOTTOM: u32 = 110;
 pub fn get_window_id(window: &Window) -> Option<u64> {
     #[allow(unused_variables)]
     match window.raw_window_handle() {
-        RawWindowHandle::Xlib(handle) => Some(handle.window),
+        RawWindowHandle::Xlib(handle) => Some(handle.window as u64),
         RawWindowHandle::Win32(handle) => Some(handle.hwnd as u64),
         RawWindowHandle::AppKit(handle) => {
             #[cfg(target_os = "macos")]
